Rewrote the StimTimer::Tick() code (WIP).
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@904 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ DarkMod/StimResponse/StimResponseCollection.cpp
@@ -595,7 +595,7 @@ void CStimResponseCollection::CreateTimer(const idDict *args, CStim *stim, int C
 		
 		// timer starts on map startup by default, otherwise wait for start
 		if( !(args->GetBool( va("sr_timer_waitforstart_%u",Counter), "0" )) ) {
-			timer->Start(sys->GetClockTicks());
+			timer->Start(static_cast<unsigned long>(sys->GetClockTicks()));
 		}
 	}
 }
